+++ include/entities/entity_model.h
@@ -148,6 +148,7 @@ public:
   // Displaying in the editor.
   virtual void draw(QPainter& painter) const;
   virtual void notify_tileset_changed(const QString& tileset_id);
+  virtual void notify_being_added();
 
   void reload_sprite();